Move label abbreviation helpers out of zoom.ts into labels.ts
abbreviateTeamName, abbreviateStageItem, and shortCourtLabel are pure
name-formatting utilities with no dependency on zoom configuration.
Extracted to logic/planning/labels.ts with a matching labels.test.ts;
zoom.ts now contains only zoom-level types, constants, and navigation.
