+++ frontend/src/components/utils/referee.tsx
@@ -3,7 +3,7 @@ import { GiWhistle } from '@react-icons/all-files/gi/GiWhistle';
 import { ReactNode } from 'react';
 
 import { MatchWithDetails } from '@openapi';
-import { abbreviateTeamName } from '@logic/planning/zoom';
+import { abbreviateTeamName } from '@logic/planning/labels';
 import { formatStageItemInput } from './stage_item_input';
 
 type RefereeFields = Pick<MatchWithDetails, 'referee' | 'referee_name'>;
